# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotelia Multi-Tenant Hotel Management API",
        "description": "Production-grade RESTful API documentation for Hotelia multi-tenant hotel backend application.",
        "contact": {
            "email": "support@hotelia.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/hotels/{hotel}/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "List all bookings for a hotel",
                "operationId": "ea4d3c107991180df79b7c4b2c272f59",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookings list retrieved"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking reservation with overlap validation",
                "operationId": "d520dbc9b43601ef7da10db904edc09d",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "guest_id",
                                    "check_in_date",
                                    "check_out_date",
                                    "room_ids"
                                ],
                                "properties": {
                                    "guest_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "check_in_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-01"
                                    },
                                    "check_out_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-05"
                                    },
                                    "room_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully"
                    },
                    "422": {
                        "description": "Validation failed or room overlap detected"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/hotels/{hotel}/bookings/{booking}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get booking details",
                "operationId": "92e5d5f9cd3da8029c6530775a536e0a",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details retrieved"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update booking details",
                "operationId": "e068c0620dc74eb7f6983c96d0a1eb9f",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking updated successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/hotels/{hotel}/bookings/{booking}/cancel": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancel a booking",
                "operationId": "4882500236d2cd8d2707b65c7283380d",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking cancelled successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/hotels/{hotel}/bookings/{booking}/check-in": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Check in guest and update rooms to occupied",
                "operationId": "94762a933d6635cb642f56471b11226c",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking checked in successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/hotels/{hotel}/bookings/{booking}/check-out": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Check out guest, update rooms to cleaning, and auto-create housekeeping tasks",
                "operationId": "42cd6f0f21f8325a4bd03b5fe6618530",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking checked out successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/hotels/{hotel}/bookings/{booking}/no-show": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Mark booking as no-show and release rooms",
                "operationId": "8f67bcfb65694fdfed37a2d2c583f5a8",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking marked as no-show successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Sanctum API token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}